[modules/blobs] setting and using logging module for logs through the variable self.logger, protecting the delete_blob_ws operation with a lock

+++ ipol_demo/modules/blobs/blobs.py
@@ -25,9 +25,12 @@ import inspect
 import string
 import re
 import ConfigParser as configparser
+import threading
+import logging
+import time
+
 from database import Database
-from error import DatabaseError, print_exception_function,\
-print_exception_thumbnail, print_exception_zip
+from error import DatabaseError
 from collections import defaultdict
 from mako.lookup import TemplateLookup
 
@@ -84,6 +87,9 @@ class   Blobs(object):
         Initialize Blob class
         """
 
+        # control the concurrent access to the blobs instance
+        self.blobs_lock = threading.Lock()
+        
         self.tmp_dir = cherrypy.config['tmp.dir']
         self.final_dir = cherrypy.config['final.dir']
         self.thumb_dir = cherrypy.config['thumbnail.dir']
@@ -98,18 +104,48 @@ class   Blobs(object):
         
         self.database_dir = "db"
         
+        self.logs_dir = cherrypy.config.get("logs_dir")
+        try:
+            if not os.path.exists(self.logs_dir):
+                os.makedirs(self.logs_dir)
+        except Exception as e:
+            self.logs_dir = os.path.dirname(os.path.realpath(__file__))
+            self.logger   = self.init_logging()
+            self.logger.exception(
+                    "Failed to create log dir (using file dir) : %s".format(e))
+        else:
+            self.logger   = self.init_logging()
+        
+        self.logger.info("----- starting blobs module -----")
+                
         try:
             self.database_dir = cherrypy.config.get("database_dir")
-        except Exception:
-          print_exception_function("failed to get database_dir config")
+        except:
+            self.logger.exception("failed to get database_dir config")
         
         self.database_name = "blob.db"
         
         try:
             database_name = cherrypy.config.get("database_name")
-        except Exception:
-          print_exception_function("failed to get database_name config")
+        except:
+            self.logger.exception("failed to get database_name config")
+            
+
+    #---------------------------------------------------------------------------
+    def init_logging(self):
+        """
+        Initialize the error logs of the module.
+        """
+        logger = logging.getLogger("blobs_log")
+        # handle all messages for the moment
+        logger.setLevel(logging.DEBUG)
+        handler = logging.FileHandler(os.path.join(self.logs_dir, 'error.log'))
+        formatter = logging.Formatter('%(asctime)s; [%(filename)s:%(lineno)s - %(funcName)20s() ] %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
+        handler.setFormatter(formatter)
         
+        logger.addHandler(handler)
+        return logger
+
         
     #---------------------------------------------------------------------------
     def instance_database(self):
@@ -121,10 +157,10 @@ class   Blobs(object):
         :rtype: Database object
         """
         try:
-            data = Database(self.database_dir, self.database_name)
+            data = Database(self.database_dir, self.database_name, self.logger)
             return data
         except DatabaseError:
-            print_exception_function("Cannot instantiate Database Object")
+            self.logger.exception("Cannot instantiate Database Object")
             sys.exit(1)
             
         
@@ -198,13 +234,9 @@ class   Blobs(object):
         :param credit:          credit blob string
         :return:                hash of current blob (dictionnary) json format
         """
-        print "*********************"
-        print "*********************"
-        print "add_blob_ws"
         cherrypy.response.headers['Content-Type'] = "application/json"
 
         blob_hash = get_hash_blob(path)
-        print blob_hash
         fileformat = file_format(path)
         #hash_tmp = -1
         dic = {}
@@ -215,10 +247,8 @@ class   Blobs(object):
             data.start_transaction()
             blobid = -1
             if not data.blob_is_in_database(blob_hash):
-              print "not in database"
-              print fileformat
+              self.logger.info( "not in database, format: "+ fileformat)
             else:
-              print "in database"
               blobid = data.blob_id(blob_hash)
             data.add_blob_in_database(demo_id, blob_hash, fileformat,
                                       ext, tag, blob_set, blob_pos_in_set,
@@ -227,9 +257,9 @@ class   Blobs(object):
             data.commit()
             dic["status"] = "OK"
         except DatabaseError:
-          print_exception_function("Cannot add item in database")
-          data.rollback()
-          dic["status"] = "KO"
+            self.logger.exception("Cannot add item in database")
+            data.rollback()
+            dic["status"] = "KO"
 
         #dic["the_hash"] = hash_tmp
 
@@ -267,7 +297,6 @@ class   Blobs(object):
         if not os.path.exists(tmp_directory):
             os.makedirs(tmp_directory)
 
-        print "blob_pos",blob_pos
         # pos in set 0 by default (if the set does not exists)
         blob_pos_in_set = 0
         
@@ -283,14 +312,10 @@ class   Blobs(object):
               for i in range(1,bs[0]['size']+1):
                   if str(bs[i]['pos_in_set'])==str(blob_pos):
                       blob_pos_in_set = blob_pos
-                      print "found"
                       break
                   else:
                       blob_maxpos = max(blob_maxpos,bs[i]['pos_in_set'])
                       blob_pos_in_set = blob_maxpos+1
-
-        print
-        print "blob_pos_in_set=", blob_pos_in_set
 
         path = create_tmp_file(blob, tmp_directory)
         data = {"demo_id": demo, "path": path, "tag": list_tag, "ext": ext, 
@@ -327,7 +352,7 @@ class   Blobs(object):
             dic["list_demos"] = data.list_of_demos()
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot have the list of demos")
+            self.logger.exception("Cannot have the list of demos")
             dic["status"] = "KO"
 
         return json.dumps(dic)
@@ -350,7 +375,7 @@ class   Blobs(object):
             dic["template_list"] = data.list_of_template()
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot have the list of templates demos")
+            self.logger.exception("Cannot have the list of templates demos")
             dic["status"] = "KO"
 
         return json.dumps(dic)
@@ -395,7 +420,7 @@ class   Blobs(object):
             data.commit()
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot update template used")
+            self.logger.exception("Cannot update template used")
             dic["status"] = "KO"
 
         return json.dumps(dic)
@@ -445,7 +470,7 @@ class   Blobs(object):
             data.commit()
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot have the list of templates demos")
+            self.logger.exception("Cannot have the list of templates demos")
             data.rollback()
             dic["status"] = "KO"
 
@@ -498,7 +523,7 @@ class   Blobs(object):
         if not os.path.exists(tmp_directory):
             os.makedirs(tmp_directory)
 
-        print "tmp_directory = ", tmp_directory
+        self.logger.info("tmp_directory = "+tmp_directory)
         path = create_tmp_file(the_archive, tmp_directory)
         self.parse_archive(ext, path, tmp_directory, the_archive.filename, demo_id)
 
@@ -519,26 +544,51 @@ class   Blobs(object):
         "OK" if not error else "KO"
         :rtype: dictionnary
         """
-        data = self.instance_database()
-        cherrypy.response.headers['Content-Type'] = "application/json"
-
         dic = {}
         dic["delete"] = ""
-        try:
-            blob_name     = data.get_blob_filename(blob_id)
-            has_no_demo   = data.delete_blob_from_demo(demo_id, blob_set, blob_id)
-            data.commit()
+        
+        # wait for the lock until timeout in seconds is reach
+        # if it can lock, locks and returns True
+        # otherwise returns False
+        def waitLock(lock,timeout):
+            current_time = start_time = time.time()
+            while current_time < start_time + timeout:
+                if lock.acquire(False):
+                    return True
+                else:
+                    cond.wait(timeout - current_time + start_time)
+                    current_time = time.time()
+            return False
+        
+        # prevent simultaneous calls
+        # try to acquire lock during 3 seconds
+        if waitLock(self.blobs_lock,3):
+        
+            data = self.instance_database()
 
-            if has_no_demo:
-                dic["delete"] = blob_name
-            dic["status"] = "OK"
+            try:
+                blob_name     = data.get_blob_filename(blob_id)
+                has_no_demo   = data.delete_blob_from_demo(demo_id, blob_set, blob_id)
+                data.commit()
+                if has_no_demo:
+                    dic["delete"] = blob_name
+                dic["status"] = "OK"
 
-        except DatabaseError:
-            print_exception_function("Cannot delete item in database")
-            data.rollback()
-            dic["status"] = "KO"
+            except DatabaseError:
+                self.logger.exception("Cannot delete item in database")
+                data.rollback()
+                dic["status"] = "KO"
 
+            finally:
+                self.blobs_lock.release()
+
+        else:
+            dic["message"] = "Failed to acquire blobs lock"
+            dic["status"]  = "KO"
+            
+        cherrypy.response.headers['Content-Type'] = "application/json"
         return json.dumps(dic)
+            
 
     @cherrypy.expose
     @cherrypy.tools.accept(media="application/json")
@@ -562,7 +612,7 @@ class   Blobs(object):
             data.commit()
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot add tag in database")
+            self.logger.exception("Cannot add tag in database")
             data.rollback()
             dic["status"] = "KO"
 
@@ -615,7 +665,7 @@ class   Blobs(object):
             data.commit()
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot delete item in database")
+            self.logger.exception("Cannot delete item in database")
             data.rollback()
             dic["status"] = "KO"
 
@@ -681,7 +731,6 @@ class   Blobs(object):
         data = {"demo_id": demo_id, "blob_set": blob_set, "blob_id": blob_id}
         res = use_web_service('/delete_blob_ws', data)
 
-        print "op_remove_blob_from_demo res=",res
         if (res["status"] == "OK" and res["delete"]):
             path_file  = os.path.join(self.current_directory, self.final_dir,  res["delete"])
             path_thumb = os.path.join(self.current_directory, self.thumb_dir, ("thumbnail_" + res["delete"]))
@@ -712,11 +761,10 @@ class   Blobs(object):
         data = self.instance_database()
         try:
             template_id = data.demo_id(template)
-            print template_id
             dic["blobs"] = data.get_blobs_of_demo(template_id)
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot access to blob from template demo")
+            self.logger.exception("Cannot access to blob from template demo")
             dic["status"] = "KO"
 
         return json.dumps(dic)
@@ -753,7 +801,7 @@ class   Blobs(object):
                 #dic["blobs"][idx] = get_new_path(dic["blobs"][idx],False)
             dic["status"]            = "OK"
         except DatabaseError:
-            print_exception_function("Cannot access to blob from demo")
+            self.logger.exception("Cannot access to blob from demo")
             dic["status"]            = "KO"
 
         return json.dumps(dic)
@@ -772,7 +820,6 @@ class   Blobs(object):
         :return: list of hash blob
         :rtype: json format list
         """
-        print "get_blobs_of_demo_ws"
         cherrypy.response.headers['Content-Type'] = "application/json"
         dic  = {}
         data = self.instance_database()
@@ -782,7 +829,7 @@ class   Blobs(object):
           dic["blobs"]        = data.get_blobs_of_demo(demo)
           dic["status"]       = "OK"
         except DatabaseError:
-            print_exception_function("Cannot access to blob from demo")
+            self.logger.exception("Cannot access to blob from demo")
             dic["status"] = "KO"
 
         return json.dumps(dic)
@@ -859,7 +906,6 @@ class   Blobs(object):
         data = {"blob_id": blob_id}
         res = use_web_service('/get_blob_ws', data)
 
-        print "return = ", res["status"]
         if res["status"] == "OK":
             b_name = res["hash"] + res["extension"]
             res["physical_location"] = os.path.join(self.current_directory,
@@ -920,7 +966,7 @@ class   Blobs(object):
             dic = data.get_blob(blob_id)
             dic["status"] = "OK"
         except DatabaseError:
-            print_exception_function("Cannot access to blob from id blob")
+            self.logger.exception("Cannot access to blob from id blob")
             dic["status"] = "KO"
 
         return json.dumps(dic)
@@ -944,7 +990,7 @@ class   Blobs(object):
         try:
             lis = data.get_tags_of_blob(blob_id)
         except DatabaseError:
-            print_exception_function("Cannot access to tag from blob")
+            self.logger.exception("Cannot access to tag from blob")
         return json.dumps(lis)
 
     #---------------------------------------------------------------------------
@@ -959,7 +1005,6 @@ class   Blobs(object):
         :return: "OK" if not error else "KO"
         :rtype: json format
         """
-        print "op_remove_demo_ws({0})".format(demo_id)
         cherrypy.response.headers['Content-Type'] = "application/json"
 
         data = self.instance_database()
@@ -971,7 +1016,7 @@ class   Blobs(object):
             
             # remove from disk unused blobs
             for blobfilename in blobfilenames_to_delete:
-                print "blobfilename to delete =",blobfilename
+                self.logger.debug("blobfilename to delete ="+blobfilename)
                 path_file  = os.path.join(self.current_directory, self.final_dir,  blobfilename)
                 path_thumb = os.path.join(self.current_directory, self.thumb_dir, ("thumbnail_" + blobfilename))
                 # process paths
@@ -979,14 +1024,14 @@ class   Blobs(object):
                 path_thumb = get_new_path(path_thumb)
                 # remove blob
                 if os.path.isfile(path_file) :  
-                    print "removing ",path_file
+                    self.logger.info("removing "+path_file)
                     os.remove(path_file)
                 if os.path.isfile(path_thumb): 
-                    print "removing ",path_thumb
+                    self.logger.info("removing "+path_thumb)
                     os.remove(path_thumb)
             
         except DatabaseError:
-            print_exception_function("Cannot delete demo")
+            self.logger.exception("Cannot delete demo")
             data.rollback()
             dic["status"] = "KO"
 
@@ -1050,19 +1095,18 @@ class   Blobs(object):
         file_dest = get_new_path(file_dest)
         name = "thumbnail_" + name
         fil_format = file_format(src)
-        print "creating ",file_dest
+        self.logger.info( "creating "+file_dest)
         try:
             if fil_format == 'image':
                 try:
                     image = PIL.Image.open(src)
                 except:
-                    print "Warning: failed to open image file"
+                    self.logger.exception( "failed to open image file")
                     return
                 image.thumbnail((256, 256))
                 image.save(file_dest)
         except IOError:
-            print_exception_thumbnail("Cannot create thumbnail",\
-                inspect.currentframe().f_code.co_name)
+            self.logger.exception("Cannot create thumbnail")
 
     #---------------------------------------------------------------------------
     def parse_archive(self, ext, path, tmp_directory, the_archive, demo_id):
@@ -1105,11 +1149,10 @@ class   Blobs(object):
                     pos = _file.find(',')
                     _file_image = _file[pos+1:]
                     _file       = _file[:pos]
-                    print "file=",_file, "_file_image=",_file_image
                     has_image_representation = True
                   if _file and _file in files:
                       title = buff.get(section, 'title')
-                      print "processing:",title
+                      self.logger.debug("processing:"+title)
                       try:
                         credit = buff.get(section, 'credit')
                       except:
@@ -1133,9 +1176,8 @@ class   Blobs(object):
                       
                       res = use_web_service('/add_tag_to_blob_ws/', data)
                       
-                      print " return = ", res["status"]
+                      self.logger.debug(" return = "+ res["status"])
                       the_hash = res["the_hash"]
-                      print the_hash
                       if the_hash != -1 and res["status"] == "OK":
                         file_dest = self.move_to_input_directory(tmp_path,
                                                                 the_hash,
@@ -1169,8 +1211,9 @@ class   Blobs(object):
                   else:
                       pass
         else:
-            print_exception_zip(inspect.currentframe().f_code.co_name,\
-                                the_archive)
+            self.logger.exception(
+                            "ZipError: index.cfg missing in " + the_archive +
+                            ": Cannot add item in database")
     
     #---------------------------------------------------------------------------
     @cherrypy.expose
@@ -1199,7 +1242,7 @@ class   Blobs(object):
             cherrypy.engine.exit()
             data["status"] = "OK"
         except Exception as ex:
-            print "something went wrong in blobs module"
+            self.logger.error("something went wrong in blobs module")
             sys.exit(1)
         return json.dumps(data)
 
@@ -1268,8 +1311,9 @@ def file_format(the_file):
       fileformat = mime.from_file(the_file)
       return fileformat[:5]
     else:
-      print "This version of magic module is not officially supported by Python"
-      print ", we advise you to install it from PIP."
+      self.logger.warning("This version of magic module is not officially "\
+                          "supported by Python, "\
+                          "we advise you to install it from PIP.")
       # if dist version
       if "open" in dir(magic):
         m = magic.open(magic.MAGIC_MIME)
